added ng-bootstrap, built barcodescanner in wrong component originally, moved to correct component
