+++ src/app/scan-bar-code/scan-bar-code.component.css
@@ -1,0 +1,20 @@
+#interactive.viewport {
+    position: relative;
+    width: 100%;
+    height: auto;
+    overflow: hidden;
+    text-align: center;
+}
+
+#interactive.viewport > canvas,
+#interactive.viewprt > video {
+    max-width: 100%;
+    width: 100%
+}
+
+canvas.drawing,
+canvas.drawingBuffer {
+    position: absolute;
+    left: 0;
+    top: 0;
+}
