--- src/app/scan-invoice/scan-invoice.component.css
@@ -1,20 +1,0 @@
-#interactive.viewport {
-    position: relative;
-    width: 100%;
-    height: auto;
-    overflow: hidden;
-    text-align: center;
-}
-
-#interactivee.viewprot > canvas,
-#interactive.viewport > video {
-    max-width: 100%;
-    width: 100%;
-}
-
-canvas.drawing,
-canvas.drawingBuffer {
-    position: absolute;
-    left: 0;
-    top: 0;
-}
