+++ src/app/scan-invoice/scan-invoice.component.html
@@ -1,27 +1,5 @@
 <h3>Scan Invoice</h3>
 <br />
 
-<div>
-    <ngb-alert type="info" [dismissible]='false'>
-        <strong>Scan</strong> the bar code
-    </ngb-alert>
-</div>
-<div *ngIf="barcode">
-    <ngb-alert type='success' [dismissable]='false'>
-        {{barcode}}
-    </ngb-alert>
-</div>
-<button type='button' class='btn btn-warning' (click)="startScanner()">
-    Start Scanning the Barcode
-</button>
-
-<div class='input-group'>
-    <div id="inputBarcode" style='position: static'>
-        <div id="interactive" class=>'viewport</div>
-    </div>
-</div>
-
-
-
 
 <h4><a href='#' routerLink='/edit-invoice'>Manually Enter Invoice</a></h4>
